@@ -1,39 +1,60 @@
 {
   "name": "liver",
-  "version": "0.1",
-  "private": true,
+  "author": "Ziga Macele",
+  "license": "MIT",
+  "version": "0.1.5",
+  "type": "module",
   "scripts": {
-    "dev": "next dev",
-    "build": "next build && next export && mv out/_next out/next && sed -i '' -e 's/\\/_next/\\.\\/next/g' out/**.html && mv out/index.html ../extension && rsync -va --delete-after out/next/ ../extension/next/ && rm -rf out && rsync -va --delete-after public/next-assets ../extension/",
-    "start": "next start",
-    "lint": "next lint"
+    "dev": "vite",
+    "build": "tsc && vite build",
+    "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
+    "lint:fix": "eslint src --ext .js,.jsx,.ts,.tsx --fix --ignore-path ./.gitignore",
+    "preview": "vite preview"
+  },
+  "lint-staged": {
+    "**/*.{js,jsx,ts,tsx}": [
+      "eslint --max-warnings=0",
+      "prettier -w"
+    ],
+    "**/*.{json,css,scss,md,webmanifest}": [
+      "prettier -w"
+    ]
   },
   "dependencies": {
-    "@heroicons/react": "^2.0.16",
-    "@next/font": "13.1.6",
-    "@types/chrome": "^0.0.217",
-    "@types/node": "18.14.0",
-    "@types/react": "18.0.28",
-    "@types/react-dom": "18.0.11",
-    "antd": "^5.4.4",
-    "axios": "^1.3.3",
-    "eslint": "8.34.0",
-    "eslint-config-next": "13.1.6",
-    "i": "^0.3.7",
-    "moment": "^2.29.4",
-    "next": "13.1.6",
-    "npm": "^9.6.5",
-    "react": "18.2.0",
-    "react-chrome-extension-router": "^1.4.0",
-    "react-dom": "18.2.0",
-    "react-fast-marquee": "^1.5.2",
-    "react-icons": "^4.7.1",
-    "react-spinners": "^0.13.8",
-    "typescript": "4.9.5"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   },
   "devDependencies": {
-    "autoprefixer": "^10.4.13",
-    "postcss": "^8.4.21",
-    "tailwindcss": "^3.2.7"
+    "@commitlint/cli": "^17.7.1",
+    "@commitlint/config-conventional": "^17.7.0",
+    "lint-staged": "^14.0.1",
+    "@crxjs/vite-plugin": "beta",
+    "@types/chrome": "^0.0.246",
+    "@types/node": "^20.6.2",
+    "@types/react": "^18.2.15",
+    "@types/react-dom": "^18.2.7",
+    "@typescript-eslint/eslint-plugin": "^6.7.2",
+    "@typescript-eslint/parser": "^6.7.2",
+    "@vitejs/plugin-react": "^4.0.3",
+    "autoprefixer": "^10.4.15",
+    "clsx": "^2.0.0",
+    "eslint": "^8.49.0",
+    "eslint-config-prettier": "^9.0.0",
+    "eslint-import-resolver-typescript": "^3.6.0",
+    "eslint-plugin-import": "^2.28.1",
+    "eslint-plugin-prettier": "^5.0.0",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "eslint-plugin-react-refresh": "^0.4.3",
+    "eslint-plugin-simple-import-sort": "^10.0.0",
+    "eslint-plugin-unused-imports": "^3.0.0",
+    "husky": "^8.0.3",
+    "postcss": "^8.4.30",
+    "prettier": "^3.0.3",
+    "prettier-plugin-tailwindcss": "^0.5.4",
+    "tailwind-merge": "^1.14.0",
+    "tailwindcss": "^3.3.3",
+    "typescript": "^5.2.2",
+    "vite": "^4.4.5",
+    "vite-tsconfig-paths": "^4.2.1"
   }
 }
